+++ files/RELEASES.md
@@ -110,3 +110,13 @@ Based on:
 - [python v4.0.1] files
 ### Releases
 - [PyPI v4.0.1] https://pypi.org/project/wingspan_files/4.0.1 - files
+
+## 2024-02-19 01:22:17
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.183.2 (2.262.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.0.2] files
+### Releases
+- [PyPI v4.0.2] https://pypi.org/project/wingspan_files/4.0.2 - files